// src/api/models/ExamQuestionEdit.ts
export interface ExamQuestionEdit {
    /**
     * 
     * @type {string}
     * @memberof ExamQuestionEdit
     */
    operationId: string;
    /**
     * 
     * @type {string}
     * @memberof ExamQuestionEdit
     */
    changeSetId: string;
    /**
     * 
     * @type {string}
     * @memberof ExamQuestionEdit
     */
    op: ExamQuestionEditOpEnum;
    /**
     * 
     * @type {string}
     * @memberof ExamQuestionEdit
     */
    clientId: string | null;
    /**
     * 
     * @type {string}
     * @memberof ExamQuestionEdit
     */
    afterClientId: string | null;
    /**
     * 
     * @type {ExamQuestionType}
     * @memberof ExamQuestionEdit
     */
    type: ExamQuestionType | null;
    /**
     * 
     * @type {string}
     * @memberof ExamQuestionEdit
     */
    yaml: string;
    /**
     * 
     * @type {NormalizedExamQuestionPayload}
     * @memberof ExamQuestionEdit
     */
    normalized?: NormalizedExamQuestionPayload;
}


/**
 * @export
 */
export const ExamQuestionEditOpEnum = {
    Update: 'update',
    Create: 'create',
    Delete: 'delete',
    Reorder: 'reorder'
} as const;
export type ExamQuestionEditOpEnum = typeof ExamQuestionEditOpEnum[keyof typeof ExamQuestionEditOpEnum];